feat/#155 :: DataLogger -> Logger 범용적인 로거로 리팩토링
